refactor(tasks): separar en componentes y paginas las tareas  Co-authored-by: 11Nataly 19nat.9295@gmail.com
